+++ VERSION_HISTORY.md
@@ -1,3 +1,6 @@
+### v5
+* run-pairsqc-single.sh available now.
+
 ### v4
 * updated cooler (Issues with pairix file input have been resolved)
 * run-cool2multirescool.sh available now.